Fix multi-job metric collection
